+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- Added more functions to support UsageReport creation in Python,
+  plus exposed the `DailyProjectUsageReport` class.
+
 ## [0.16.0] - 2025-08-08
 
 ### Added